+++ zarrs/src/array/array_async_readable_writable.rs
@@ -158,15 +158,13 @@ impl<TStorage: ?Sized + AsyncReadableWritableStorageTraits + 'static> Array<TSto
                 .await?;
 
             // Update the chunk
-            let chunk_bytes_new = unsafe {
-                update_array_bytes(
-                    chunk_bytes_old,
-                    &chunk_shape,
-                    chunk_subset,
-                    &chunk_subset_bytes,
-                    self.data_type().size(),
-                )
-            };
+            let chunk_bytes_new = update_array_bytes(
+                chunk_bytes_old,
+                &chunk_shape,
+                chunk_subset,
+                &chunk_subset_bytes,
+                self.data_type().size(),
+            )?;
 
             // Store the updated chunk
             self.async_store_chunk_opt(chunk_indices, chunk_bytes_new, options)