+++ CHANGELOG.md
@@ -10,12 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `array:codec::{InvalidBytesLengthError,InvalidArrayShapeError,InvalidNumberOfElementsError,SubsetOutOfBoundsError}`
 - Add `ArraySubset::inbounds_shape()` (matches the old `ArraySubset::inbounds` behaviour)
+- Add `ArrayBytesFixedDisjointView[CreateError]`
 
 ### Changed
 - **Breaking**: change `ArraySubset::inbounds` to take another subset rather than a shape
 - **Breaking**: `CodecError` enum changes:
   - Change `CodecError::UnexpectedChunkDecodedSize` to an `InvalidBytesLengthError`
   - Add `CodecError::{InvalidArrayShape,InvalidNumberOfElements,SubsetOutOfBounds}`
+- **Breaking**: Change output args to `ArrayBytesFixedDisjointView` and make safe the following:
+  - `Array::[async_]retrieve_chunk[_subset]_into`
+  - `[Async]ArrayPartialDecoderTraits::partial_decode_into`
+  - `ArrayToBytesCodecTraits::decode_into`
+  - `zarrs::array::copy_fill_value_into`
+  - `zarrs::array::update_array_bytes`
 
 ## [0.19.1] - 2025-01-19